+++ app/Http/Livewire/Post/LikeButton.php
@@ -4,14 +4,15 @@ namespace App\Http\Livewire\Post;
 
 use App\Actions\Post\CreateLikePost;
 use App\Actions\Post\DeleteLikePost;
+use App\Models\Post;
 use Livewire\Component;
 use App\Models\Like;
 use Illuminate\Support\Facades\Auth;
 
 class LikeButton extends Component
 {
-    public $post;
-    public $counts = 0;
+    public Post $post;
+    public int $counts = 0;
 
     public function mount(): void
     {
