+++ resources/views/layouts/app.blade.php
@@ -28,6 +28,7 @@
     </head>
     <body class="font-sans antialiased">
         <div class="min-h-screen bg-gray-100">
+            <x-jet-banner/>
             @livewire('navigation-menu')
             <!-- Page Heading -->
             <header class="bg-white shadow">
