+++ resources/views/livewire/post/delete-post-button.blade.php
@@ -1,0 +1,5 @@
+@if ($post->authorIs(auth()->user()))
+    <x-jet-danger-button wire:click="deletePost" class="ml-2">
+        {{ __('削除する') }}
+    </x-jet-danger-button>
+@endif
