+++ resources/views/post/show.blade.php
@@ -10,7 +10,10 @@
                         {{ $post->title }}
                     </p>
                 </div>
+
                 @livewire('stuff.permit-button', ['post' => $post])
+
+                @livewire('post.delete-post-button', ['post' => $post])
             </div>
         </h2>
     </x-slot>
@@ -20,10 +23,16 @@
             <div class="flex justify-between container mx-auto">
                 <div class="w-full">
                     @if (session()->has('message'))
-                        <div class="relative flex flex-col sm:flex-row sm:items-center bg-white shadow rounded-md py-5 pl-6 pr-8 sm:pr-6 mb-5">
-                            <div class="flex flex-row items-center border-b sm:border-b-0 w-full sm:w-auto pb-4 sm:pb-0">
+                        <div
+                            class="relative flex flex-col sm:flex-row sm:items-center bg-white shadow rounded-md py-5 pl-6 pr-8 sm:pr-6 mb-5">
+                            <div
+                                class="flex flex-row items-center border-b sm:border-b-0 w-full sm:w-auto pb-4 sm:pb-0">
                                 <div class="text-green-500">
-                                    <svg class="w-6 sm:w-5 h-6 sm:h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M9 12l2 2 4-4m6 2a9 9 0 11-18 0 9 9 0 0118 0z"></path></svg>
+                                    <svg class="w-6 sm:w-5 h-6 sm:h-5" fill="none" stroke="currentColor"
+                                         viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
+                                        <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2"
+                                              d="M9 12l2 2 4-4m6 2a9 9 0 11-18 0 9 9 0 0118 0z"></path>
+                                    </svg>
                                 </div>
                                 <div class="text-base ml-3">
                                     Awesome 👍
@@ -68,7 +77,7 @@
                                         <span class="mx-2 md:mx-0">
                                             @livewire('post.like-button', ['post' => $post])
                                         </span>
-                                        <x-user-avatar :user="$post->user" />
+                                        <x-user-avatar :user="$post->user"/>
                                         <span class=" font-light text-gray-600">
                                             {{ $post->created }}
                                         </span>
